C: Removing from Platforms class the random generator initialization. CPP: fixing warning
git-svn-id: https://simple-tetris-clone.googlecode.com/svn/trunk@61 41f6d736-5e53-0410-a2ad-4df87bb8d04b

+++ src/sdl/sdl_game.cpp
@@ -20,7 +20,7 @@ namespace Stc {
 int PlatformSdl::init(Game *game)
 {
     // Initialize the random number generator
-    srand(time(NULL));
+    srand((unsigned int)(time(NULL)));
 
     // Start video system
     if (SDL_Init(SDL_INIT_VIDEO) < 0)
